requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
backports.lzma==0.0.14
beautifulsoup4==4.12.3
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.4
deep-translator==1.11.4
distlib==0.3.8
easyocr==1.7.1
exceptiongroup==1.2.0
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.13.3
fsspec==2024.3.1
future==1.0.0
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.35
idna==3.6
imageio==2.34.0
importlib_resources==6.4.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
nodeenv==1.8.0
numpy==1.24.4
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
pre-commit==3.5.0
pyclipper==1.3.0.post5
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
pyTelegramBotAPI==4.16.1
python-bidi==0.4.2
python-telegram-bot==21.0.1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.3
sympy==1.12
telebot==0.0.5
tifffile==2023.7.10
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.25.1
zipp==3.18.1
